2023-06-25: After spending too much time dreaming and writing-out ideas in Zim, I'm finally creating the first "housing" structure in zim.
In this first commit I've added a spreadsheet that determines the width and height and number of bedrooms. The idea is that it'll be shaped rectangular like an apartment. I don't want to make it wider because I want every room to have balcony access. The structure will grow to 6 stories, and if that's not enough, I'll add 3x more around a courtyard. Then that 4x design can be duplicated to form blocks and streets, if needed.

Anyway, all I have now are the 4 external walls and a hallway. Now I need to add some room dividers. Then I want to experiment with the subsequent floors and the balcony.