made the 5th floor and the terraces going-up slightly smaller
I'm thinking that the actual roof of every terrace will be equal, but the floor will end shorter as it gets higher. The floor will be extended (after a railing) by transparent roofing. This will allow more light into the lower terraces without getting rain on the edges (where we may like to place tables & chairs)