added last triangle, but it's shit
I couldn't figure out how to draw a sketch in three dimensions. What I really need is to just move the vertex where all the triangeles meet up by, say, 0.5 meters. That's it. But all I could do was draw another sketch for each triange, pad it into a 100 mm part, and then use the "transform" GUI tool to eyeball it. Well, that made the last piece horribly out-of-whack.

How the heck do I just modify the z-coordinates of one vertex in a sketch on the xy plane?