ok, holy shit, I figured out how to sketch in 3 dimensions..sort-of
This commit adds a sloped plane that runs from the previously flat rooftop floor up to a ridgeline that was floating 1 meters above it, straddling the center.

Well, it's just a triangle for now because it was easier than an arch. I'll do the arch next.

Basically the way this worked is that I took my existing sketch of the roottop floorplan on the xy plane, then I created a *new* sketch that hovered a single ridgeline *above* that one, also on the xy plane. Note that you can tie a new sketch to another sketch by using the "Create an edge linked to an existing geometry" button.

At this point, the new sketch on the xy plane had just one line sitting at the origin. SO I just changed the z-position of the new sketch so that it was 1 meter above the roof floor sketch.

Finally, I switched to draft tool and (using the end-line-snap), I was able to sketch this new plane that does exactly what I want across all three dimensions. Awesome!