added the 16 square meter Scheffler concentrator. The size of the dish in freecad is 5 meters in diameter, but it should actually be 5 meters in one axis and 4 meters in another axis. So they're a bit larger then they will actually be, but they really are *big*.
Obviously the focal point will be aimed at a collection chamber, which is absent. But this roughly shows us that we can fit three dishes. They'll shade the PV rays, but I don't see a way around that -- other than putting PV on one rooftop and solar thermal on another (once the second building is built)