added a single stair
after some research, it looks like the Irish government defines the "optimum" emergency (fire) egress stairwells of having a width of 1 meter, a rise of 150 mm, and a going/run of 300mm.

 * https://www.gov.ie/pdf/?file=https://assets.gov.ie/100533/6926c022-4a80-4a05-8a55-dda6347bdaca.pdf#page=null
 * https://www.gov.scot/publications/building-standards-technical-handbook-2020-non-domestic/2-fire/2-9-escape/

There should be two stairwells for each building (which will be used to access two buildings each). The stairswells will be fully enclosed (no windows) exterior structures (fire-safe) that connect to the terraces of each floor. They will go from the basement to the 6th floor terrace (a ladder with a locked hatch (openable with a key from below and without a key from above) will go to the rooftop.

There should be overhead LED emergency lights with independent batteries that turn-on when [a] motion is detected and [b] when power is cut. The batteries should supply enough power to illuminate for ~3 hours.

Eventually, a machine room will be installed on-top of the stairwell that will power an elevator that will climb along the exterior of the stairwell.

Every floor of the stairwell should have 3 doors:
1. One door goes to the first Launch Nest building
2. One door goes to the adjacent/perpendicular Launch Nest building
3. One door goes out to the elevator

I'd want each flight to go up one-half-building-floor with a landing connecting the turns. Every other landing will have the three doors with an L-shape landing that wraps around the side

TODO: build the other steps to complete the stairwell