ok, now I have the stairs perfect
The problem was that I was setting the "tread depth enforce" to 300 mm. When I used the ruler, I saw that the "rise" was some fraction above 150 mm or something. It wasn't even and it was fucking everything up.

I don't see a way to set the "going" to 300 mm, but it appears to determine that based on the rise and the line that you've drawn. So I moved the landing to where I expected it mathmatically (-500+2900=2400). That was ok, but off-by-one (I guess because we're using the landing itself as a step to save space). So I put it to 2100 m in the x-plane, drew a line from the top of the bottom landing (on-center) to the bottom of the next landing up (on-center), set the rise to 150 mm, and used the ruler tool to confirm the "going" of the steps are 300 mm. Perfect