created a clone of the second-floor terrace sketch, so now there's one "high" terrace sketch and one "low" terrace sketch. The difference is the z-plane.
I decided to have a 3-5 degree pitch on all terraces. In the case of the second-floor terrace, it extends 8 meters out from the exterior walls + 3.5 meters across the short-side of the structure = 11.5 meters total run.

If the rise (fall) of the terrace is 0.7 meters across that 11.5 meter run, it's a 3.483 degree slope that meets our requirements

 * https://www.omnicalculator.com/construction/roof-pitch

Having a pitched slope entering a doorway is a problem, so to make a ramp with no steps across the doorway's entrance, I decided to drop the entire terrace down 4 cm. This allows us to put a 0.5 meter ramp up to the doorway with a 0.08 rise/run. The sides of the ramp will essentially end-up being a 4 cm "step", but that's OK.