added a surface to the now-sloping second-floor terrace.
Once I had the two sketches of the terrace (one at the "top" edge of the terrace and one at the "bottom" edge of where the terrace would slope-down to), I added a third copy of the terrace sketch to be exactly between the high & low skteches.

This then allowed me to use the "Arc by 3 points" tool in the Draft Workbench to draw the arcs sloping from the top sketch to the middle sketch to the low bottom sketch.. But rather than join/upgrade these, I just hopped over to the Surface Workbench and clicked Surface -> Fill Boundary Curves.

Then I clicked "Add Edge" and then selected all four of the sloped arcs, and it did exactly what I needed (created a surface/plane along those edges). I clicked "OK".

For more information on this, see my post on the FreeCAD forums:

 * https://forum.freecad.org/viewtopic.php?p=700353#p700353