added interior walls to the 4th floor, which includes the 4x 3meter-by-3meter bedrooms and the common space in-between
TODO: divide the common-space to add rooms for:

 1. Telecommunications & Server Room
 2. Storage Room
 3. Small Urine WC
 4. Laundry Room

The Server Room needs to be secure, insulated, and Air Conditioned.

The Storage Room should be secure, so it can be used to store some expensive items, if needed.

The laundry room doesn't need to be secured, but we should put it in a thick-walled room with heavy doors to contain noise

There will also been a "locker room," but this doesn't need to be in a room at all. Actually, it's better to have this in the open