Added bedroom door
I started working on the bedroom interior doors, but I realized there's four of them all in-a-row. So, instead of fighting with 4 distinct window elements, I just added one with a "depth" of 99 meters. That can be cleaned-up later, if needed.

However, the door to the terrace on one of the bedrooms seems obstructed somehow. I saw this issue also when I was making individual doors. Not sure what's causing it.