added sloped terrace to floor #4
While the terrace extended 8 meters-out from the building's exterior walls on the 2nd floor, the 4th floor terrace extends-out only 7 meters. Instead of dropping 0.7 meters over 8 + (7/2) = 11.5 meters of run (3.483 degree pitch = 6.087 % grade), we're going to drop 0.6 meters over 7 + (7/2) = 10.5 meters of run (3.2705 degree pitch = 5.714% grade)

This round number (0.6 meters) satisfies our requirements of being between 3-5 degrees of slope. We want at least 3 degrees so that it'll drain when spraying-down the terrace with water, and we also want it to be as low (close to 3.00) as-is reasonable to minimize the annoyance of a non-level floor