Updated TODO to include an event space in the basement
While the 6th floor terrace is great for some events, I realized that some other types of events (eg parties with loud music) wouldn't be so nice for neighbors to be blasting noise from a rooftop high above the tree lines.

Rather, let's reduce the store rooms in the basement from 3m x 6m down to 3m x 3m and replace two of the rooms with just one large open space. That'll give us 15m x 7 m = 105 square meters.

According to the "4 sq meter rule", that will give us a capacity of 26.25 people. Our structure is built to sleep 30, so that's perfect. In fact, the "2-4 sq meter rule" is for comfortable corporate/private events. A dance party should still be able to safely squeeze a lot more people-in -- especially if people that are *not* dancing can go hang-out on the first floor.