added "_simple" freecad, which has 5 floors only (instead of 6)
after reading Mexican, Colombian, and EU building codes for confined masonry structures, it became clear that 5 stories is the max. There are some others, eg India, that allow 6-stories -- but probably it would be best to be conservative here and stick with the most restrictive building codes, especially because many consumers of our designs will not be professional civil engineers or builder.

According to the Eurocode, these "storey" restrictions do not restrict below-ground floors (eg basements), so 5 storeys above ground + 1 storey below ground should be OK.

 * page 203 https://www.confinedmasonry.org/wp-content/uploads/2009/09/Eurocode-8-1-Earthquakes-general.pdf

I'm also going to have to drastically reduce the size of the windows. According to the Peruvian building code, "window openings must be less than half the length of the wall".

 * page 15 https://www.confinedmasonry.org/wp-content/uploads/2009/09/Peru_Booklet-masonry-v03.pdf)

Also, according to Eurocode 8, walls with windows and doors cannot contribute to the wall density calculations. So I'm going to need to make more walls without windows :(

 * page 40 https://www.confinedmasonry.org/wp-content/uploads/2009/09/ConfinedMasonryDesignGuide82011.pdf

I also may need to increase the length of the sheer wall to 5.4 meters, which is 30% of the length of the entire structure, as required by the EU regs.