added pressurized water tank to rooftop
This commit adds a note to the README that the rooftop must contain a pressurized water tank.

Solar equipment (PV panels, parabolic solar concentrators) work best when they can absorb or reflect as much light as possible. Over time, dust and dirt and volcanic ash will accumulate on this equipment, reducing their efficacy. The solution is simple: a few times per year, wash them with water.

To minimize shade and maximize energy collection, we put all our water tanks on the top floor, a few meters below the rooftop. Without a way to pump water up to the rooftop, someone would have to haul buckets of water up there. That's not nice, and it wouldn't work very well for cleaning a 16 square meter parabolic dish. We really need a hose with pressure to properly wash all of the equipment.

While I generally prefer all water pressure to be generated by gravity, cleaning the rooftop equipment is done so infrequently (only a few times per year), that we can tolerate only doing it on a sunny day.

So we'll just put a very small pressurized water tank on the rooftop with an electric pump. This can pump the water up from the tank on the terrace floor below it into and generate enough pressure for a hose to spray out to all our PV panels from the aisles and up to the top of the parabolic dishes' top mirrors.