TODO: change height of each storey
This commit updates the todo.txt file with some details about why we need to change the storey height.

Currently we have 2.4 meters between each floor on-center, but that's with a 0-mm-thick concrete slab between each storey's walls. In reality, we need 2.4 meters of clearance between the top of the floor and the lowest point of the ceiling above (so including light fixtures, duct work, etc)

But, since we set the height to 2.4 meters, I've also learned more about the sound transfer between floors. Previously I thought we could just have a very thick concrete slab, but it turns out that that won't prevent the transfer of impact noise between floors. So you need a reasonably thick slab (~150-200 mm) *and* a decoupled floor above it that sits on resilient pads (there's other ways, but this seems like the best option; I don't want carpets)

 * https://engineering.stackexchange.com/questions/60671/how-to-build-apartment-floors-ceilings-to-not-transfer-sound/60672#60672

Probably the easiest safe option is to just go with 3-meters between storeys. The best option is to calculate how thick the slabs should be to support the building loads (with an engineer), figure out the maximum size of the ductwork that will run along the ceiling, and figure out the height of the floating floor off the concrete slab. Then we can sum that together with 2.4 m between, and we'll get the height of each storey.