adding an svg "urban plan" file to map out how many launch nests fit together in an ideal landscape
In this commit, I've created a file 'urban_plan.svg'. It's 2,000 cm x 2,000 cm. Each cm is supposed to represent 1 meter in real life.

The launch nest has an external wall width of 18 meters, but when you include the terrace, it's 25 meters long. That doesn't include the stairwell, which adds another 3 meters.

If we arrange 4 launch nests into a + shape, the width becomes <staircase> + <terrace> + <staircase> + <terrace> + <staircase>. That's 3 + 25 + 3 + 25 + 3 = 56 meters. But the actual structure will likely be longer, since we didn't account for the wall thickness in our original CAD design (TODO after we talk with an engineer and get the actual wall thicknesses). So, probably we should just make it 60 meter x 60 meters or something.

So this vector drawing has areas set-aside for construction in black.

	TODO: I accidentally made this 33 meters for the Launch Nest. I need to fix it to ~60 meters.

Between each Launch Nest is a Treasure Tower. Our latest edition of the treasure tower (from a few months ago) has a max width in one direction of 15 meters. But, again, it may increase due to wall thickness (I do really want to meet the CP min dimensions of 4x3 meters and have two on the ground-floor). So I made this area 18 meters x 18 meters.

Around each of these black "structure areas", I've drawn a circle that's 30 cm from the edge of the structure area. This is our "short tree" area, as I don't want any tall (>30 meter) trees growing within 30 meters of the structure. The idea is that trees get placed X distance from the structure, such that X is the max growing height for that tree. So this light-green "small tree area" will have trees increasing in heigh as you get further from the actual structures. But, basically, this "light green area" is our orchard area.

In the center of a square of four of these arrangements is a big dark green space. This is the "conservation" area. It's going to be about 1 hectare. The idea is that this area is fenced-off. It should be native, wild jungle. We don't build paths through it. We don't cut anything down in it (except maybe non-native invasives). This area is for the wild animals.

The background is another green. This is our "large trees" area.

What's not shown is the expectation that we'll leave a buffer area around streams at least 10 meters -- definitely >=30 meters for structures).

Obviously, land isn't flat and there's rivers running though it. So this isn't going to be practical. But it's a useful guide to understand the minimum amount of land required to build X number of launch nest quads.