fixed the sizes of the urban plan
This commit changes the size of the "construction area" for the launch nest quads to be 64 meters x 64 meters, due to a logic error in my last commit's haste.

I also increased the size of the Treasure Tower to 30 meters. This is a ton, but it would allow us to actually squeeze 4 treasure towers into the space, if needed. Then, if we needed to scale, we wouldn't have to destroy the whole tower -- we'd just maybe need to demolish the ramps to squeeze another tower next to it

I also added a 2-meter-wide pathway between all of the launch nests. Because this pathway will be covered (so folks can walk to the toilets in the rain), we expand the "zone 2" area (with short trees only) to either side of this pathway.

I also added a (bicycle) road on the inside. Currently it has sharp edges, though in reality the corners would be rounded.

TODO: add a "zone 3" background buffer behind the road. This will double as a fire break between the heavily-forested zones 4 & 5 and our orchard in zone 2.