added zone 3
This commit adds the permaculture "zone 3". This is basically just our (bicycle) road, but it also serves as a fire buffer. I made it 12 meters wide, which includes a 4-meter-wide road in the middle

Currently it overlaps the zone 6 conservation zone. In practice I think we'd actually have this road be straight and instead only curve to meet high-velocity highway road specifications at the edges, so that wouldn't happen. But for now a circle is easier to draw.

TODO: draw the buildings (zone 0) and the intensive gardens in the building courtyards (zone 1)