added zones and key to urban plan
this commit adds 4x launch nests to each "quad". I also added a legend indicating which colors are which permaculture zones.

Finally, we have the total dimensions. For a 4-quad urban plan on an ideal landscape, the width is 338 meters x height 338 meters = 114,244 sq meters = 11.4 hectares. I was expecting it to be closer to 20 or 30, so that's not bad :)